+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Added throttling for mouse wheel scroll events when using custom mouse wheel scrolling as an attempt to make the wheel scrolling smooth
+
+- Specify in the example app that it is also possible to use HOME and END keys for keyboard scrolling
+
 ## 0.0.1
 
 - Initial release.